pyproject.toml
[project]
name = "toon-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.5",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.2",
    "python-dotenv>=1.0.0",
    "python-toon>=0.1.3",
    "toon-format>=0.1.0",
]

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.4.7",
]

[tool.uv.sources]
toon-format = { git = "https://github.com/toon-format/toon-python.git", rev = "main" }
